Moved reproduction functions to epymodelingsuite.calibration
